requirements.txt
openai>=1.0.0
Pillow>=10.0.0
rich>=13.0.0
emoji>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
